Test: 테스트 코드 메서드 정리
