+++ src/main/java/com/kkiri_trip/back/api/controller/ScheduleItemController.java
@@ -6,10 +6,7 @@ import com.kkiri_trip.back.global.common.dto.ApiResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
@@ -23,13 +20,13 @@ public class ScheduleItemController {
     @GetMapping("feeds/{feedId}/scheduleItems")
     ResponseEntity<ApiResponseDto<List<ScheduleItemDto>>> getAllScheduleItemsInFeed(@PathVariable("feedId") Long feedId)
     {
-        return ApiResponseDto.from(HttpStatus.OK, "피드의 스케줄 아이템 조회", scheduleItemService.getAllScheduleItemsInFeed(feedId));
+        return ApiResponseDto.from(HttpStatus.OK, "피드의 스케줄 아이템 리스트 조회", scheduleItemService.getAllScheduleItemsInFeed(feedId));
     }
 
     @GetMapping("/schedules/{scheduleId}/scheduleItems")
     ResponseEntity<ApiResponseDto<List<ScheduleItemDto>>> getAllScheduleItemsInSchedule(@PathVariable("scheduleId") Long scheduleId)
     {
-        return ApiResponseDto.from(HttpStatus.OK, "피드의 스케줄 아이템 조회", scheduleItemService.getAllScheduleItemsInSchedule(scheduleId));
+        return ApiResponseDto.from(HttpStatus.OK, "스케줄의 스케줄 아이템 리스트 조회", scheduleItemService.getAllScheduleItemsInSchedule(scheduleId));
     }
 
     @GetMapping("/feeds/{feedId}/schedules/{scheduleId}/items/{id}")
@@ -37,12 +34,24 @@ public class ScheduleItemController {
                                                                           @PathVariable("scheduleId") Long scheduleId,
                                                                           @PathVariable("id") Long id) {
         return ApiResponseDto.from(HttpStatus.OK,
-                "피드의 스케줄 아이템 조회",
+                "스케줄 아이템 조회",
                 scheduleItemService.getScheduleItem(
                         feedId,
                         scheduleId,
                         id
-                ));
+                )
+        );
+    }
+
+
+    @PostMapping("/feeds/{feedId}/schedules/{scheduleId}/items")
+    ResponseEntity<ApiResponseDto<ScheduleItemDto>> createScheduleItem(@PathVariable("feedId") Long feedId,
+                                                                    @PathVariable("scheduleId") Long scheduleId,
+                                                                    @RequestBody ScheduleItemDto scheduleItemDto) {
+        return ApiResponseDto.from(HttpStatus.CREATED,
+                "스케줄 아이템 조회",
+            scheduleItemService.createScheduleItem(feedId,scheduleId,scheduleItemDto)
+        );
     }
