+++ src/main/java/com/kkiri_trip/back/domain/scheduleItem/service/ScheduleItemService.java
@@ -1,8 +1,10 @@
 package com.kkiri_trip.back.domain.scheduleItem.service;
 
 import com.kkiri_trip.back.api.dto.Schedule.ScheduleItem.ScheduleItemDto;
+import com.kkiri_trip.back.domain.feed.entity.Feed;
 import com.kkiri_trip.back.domain.feed.repository.FeedRepository;
 import com.kkiri_trip.back.domain.feed.service.FeedService;
+import com.kkiri_trip.back.domain.schedule.entity.Schedule;
 import com.kkiri_trip.back.domain.schedule.repository.ScheduleRepository;
 import com.kkiri_trip.back.domain.scheduleItem.entity.ScheduleItem;
 import com.kkiri_trip.back.domain.scheduleItem.repository.ScheduleItemRepository;
@@ -60,5 +62,28 @@ public class ScheduleItemService {
                 new ScheduleItemException(ScheduleItemErrorCode.SCHEDULEITEM_NOT_FOUND))
                 .toDto();
 
+    }
+
+    public ScheduleItemDto createScheduleItem(Long feedId, Long scheduleId, ScheduleItemDto scheduleItemDto) {
+
+        feedRepository.findById(feedId).orElseThrow(()->
+                new FeedException(FeedErrorCode.FEED_NOT_FOUND));
+
+        Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(()->
+                new ScheduleException(ScheduleErrorCode.SCHEDULE_NOT_FOUND));
+
+        validateScheduleDto(scheduleItemDto.getItemOrder());
+
+        ScheduleItem scheduleItem = ScheduleItem.builder()
+                        .itemOrder(scheduleItemDto.getItemOrder())
+                        .schedule(schedule)
+                        .build();
+        return scheduleItemRepository.save(scheduleItem).toDto();
+    }
+
+    void validateScheduleDto(int itemOrder)
+    {
+        if (itemOrder <= 0 )
+            throw new ScheduleItemException(ScheduleItemErrorCode.INVALID_ITEMORDER);
     }
 }
